Move some more tests out of root functional folder
