+++ awx/main/tests/functional/api/test_api_generics.py
